% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R, R/get_option.R
\name{get_option}
\alias{get_option}
\alias{get_option}
\title{Get package options
From 'redis.progress' options() entry}
\usage{
get_option(name = NULL)

get_option(name = NULL)
}
\arguments{
\item{name}{option name to get, if null returns all options}
}
\description{
Get package options
From 'redis.progress' options() entry

Get package options
From 'redis.progress' options() entry
}
