% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{publish_queue}
\alias{publish_queue}
\title{Defined a way to publish a queue name}
\usage{
publish_queue(name, type = c("key", "list"), prefix = FALSE)
}
\arguments{
\item{name}{key name to use to publish last generated queue name}

\item{type}{type of storage to use (see details)}

\item{prefix}{use queue prefix to the key (by default false)}
}
\description{
In some context you need to create a random queue name to make it unique. By doing this you don't know
the real queue name before to run the script.
The publish feature allows to store this queue name once it's created under a predefined key name.
}
\details{
This function should be used to set the `publish` parameter of \code{\link{create_redis_progress}}


Using 'key' as type will only store the last created, 'list' will store it in a list as the last element
}
\examples{

publish_queue('my-jobs', 'list') # Store task queue names in a list under the 'my-jobs' names

}
