% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{redis_cleanup_progress}
\alias{redis_cleanup_progress}
\title{Cleanup all job queues with the given name}
\usage{
redis_cleanup_progress(name, redis = NULL)
}
\arguments{
\item{name}{name of the queue}

\item{redis}{redis client object \code{redis_client}}
}
\description{
Cleanup all job queues with the given name
}
