% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{redis_progress_monitor}
\alias{redis_progress_monitor}
\title{Monitor Progress of a tasks queue}
\usage{
redis_progress_monitor(from, redis = NULL, options = list())
}
\arguments{
\item{from}{source of the queue to monitor, simplest is the queue name (without the queue prefix)}

\item{redis}{redis client definintion object, returned by \code{redis_client()}}
}
\description{
Basic monitoring of task queue progress created using
}
