% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{RedisClientMock-class}
\alias{RedisClientMock-class}
\alias{RedisClientMock}
\title{Client interface for mocking Redis Database}
\description{
Client interface for mocking Redis Database
}
\section{Fields}{

\describe{
\item{\code{data}}{environment where data are stored}
}}

\section{Methods}{

\describe{
\item{\code{hashIncrBy(key, field, value)}}{increment [field] in [key] hash}

\item{\code{hashSetCounter(key, field, value)}}{Set counter [field] of hash named by [key]}
}}

\seealso{
\link{redis_client}

Other backend: 
\code{\link{RedisClientRRedis-class}},
\code{\link{RedisClientRcpp-class}},
\code{\link{RedisClientRedux-class}},
\code{\link{redis_client}()}
}
\concept{backend}
\keyword{internal}
