% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{RedisClientRcpp-class}
\alias{RedisClientRcpp-class}
\alias{RedisClientRcpp}
\title{Client interface for RcppRedis library}
\description{
Client interface for RcppRedis library
}
\section{Methods}{

\describe{
\item{\code{connect()}}{Connect to the Redis database}

\item{\code{delete(key)}}{Delete key}

\item{\code{exists(key)}}{Key exists
@returns boolean}

\item{\code{get(key)}}{Get key value}

\item{\code{hashGet(key, field)}}{Get field value in hash named by key}

\item{\code{hashIncrBy(key, field, value)}}{increment [field] in [key] hash}

\item{\code{hashSet(key, field, value)}}{Set [field] of hash named by [key]}

\item{\code{pushTail(key, value)}}{Push value on tail of the list named by key}

\item{\code{set(key, value)}}{Set key value}
}}

