% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{RedisClientRedux-class}
\alias{RedisClientRedux-class}
\alias{RedisClientRedux}
\title{Client interface for redux & rrlite library}
\description{
Client interface for redux & rrlite library
}
\section{Methods}{

\describe{
\item{\code{hashIncrBy(key, field, value)}}{increment [field] in [key] hash}
}}

