% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{RedisClientRRedis-class}
\alias{RedisClientRRedis-class}
\alias{RedisClientRRedis}
\title{Client interface for rredis library
Internal class.}
\description{
Client interface for rredis library
Internal class.
}
\section{Methods}{

\describe{
\item{\code{connect()}}{connect to the redis database}

\item{\code{hashGet(key, field)}}{Get the value of a field in a 'HSet'}

\item{\code{hashGetAll(key)}}{Get all fields of an 'HSet'}

\item{\code{hashSet(key, field, value)}}{set the field in a 'HSet' with a value}

\item{\code{initialize(host = "localhost", port = 6379, database = NA, ...)}}{initialize the instance with connection parameters}

\item{\code{name()}}{get the instance name}
}}

\seealso{
\code{\link{redis_client}()}

Other backend: 
\code{\link{RedisClientMock-class}},
\code{\link{RedisClientRcpp-class}},
\code{\link{RedisClientRedux-class}},
\code{\link{redis_client}()}
}
\concept{backend}
\keyword{internal}
