% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R, R/redisProgress.R
\docType{package}
\name{redisProgress}
\alias{redisProgress}
\alias{redisProgress-package}
\alias{redisProgress}
\alias{redisProgress-package}
\alias{redisProgress-package}
\title{Progress bar for distributed task with redis backend}
\description{
redisProgress Package

redisProgress Package
}
\author{
Clement Turbelin

Clement Turbelin
}
\keyword{internal}
