% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{redis_queue_name}
\alias{redis_queue_name}
\title{Get the full name of the queue including current queue's prefix}
\usage{
redis_queue_name(name)
}
\arguments{
\item{name}{name of the queue (without prefix)}
}
\description{
Get the full name of the queue including current queue's prefix
}
