% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{RedisClientRcpp-class}
\alias{RedisClientRcpp-class}
\alias{RedisClientRcpp}
\title{Client interface for RcppRedis library
Internal class}
\description{
Client interface for RcppRedis library
Internal class
}
\section{Fields}{

\describe{
\item{\code{args}}{arguments to use to connect to the redis database}

\item{\code{cnx}}{connexion instance return by the backend}

\item{\code{database}}{database number to connect to}
}}

\section{Methods}{

\describe{
\item{\code{connect()}}{Connect to the Redis database}

\item{\code{delete(key)}}{Delete key}

\item{\code{exists(key)}}{Key exists
@return boolean}

\item{\code{get(key)}}{Get key value}

\item{\code{hashGet(key, field)}}{Get field value in hash named by key}

\item{\code{hashIncrBy(key, field, value)}}{increment [field] in [key] hash}

\item{\code{hashSet(key, field, value)}}{Set [field] of hash named by [key]}

\item{\code{hashSetCounter(key, field, value)}}{Set [field] of hash named by [key]}

\item{\code{pushTail(key, value)}}{Push value on tail of the list named by key}

\item{\code{set(key, value)}}{Set key value}
}}

\seealso{
\link{redis_client}

Other backend: 
\code{\link{RedisClientMock-class}},
\code{\link{RedisClientRRedis-class}},
\code{\link{RedisClientRedux-class}},
\code{\link{redis_client}()}
}
\concept{backend}
