% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\docType{class}
\name{RedisClientRedux-class}
\alias{RedisClientRedux-class}
\alias{RedisClientRedux}
\title{Client interface for redux & rrlite library}
\description{
Client interface for redux & rrlite library
}
\section{Fields}{

\describe{
\item{\code{type}}{type of library to use  redux or rrlite (default is redux)}
}}

\section{Methods}{

\describe{
\item{\code{hashIncrBy(key, field, value)}}{increment [field] in [key] hash}

\item{\code{hashSetCounter(key, field, value)}}{Set counter [field] of hash named by [key]}
}}

\seealso{
\link{redis_client}

Other backend: 
\code{\link{RedisClientMock-class}},
\code{\link{RedisClientRRedis-class}},
\code{\link{RedisClientRcpp-class}},
\code{\link{redis_client}()}
}
\concept{backend}
\keyword{internal}
