% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redisProgress.R
\docType{package}
\name{redisProgress}
\alias{redisProgress}
\alias{redisProgress-package}
\title{Progress bar for distributed tasks with redis backend}
\description{
redisProgress Package
}
\details{
The goal of redisProgress is to create a progress bar system to follow tasks distributed across several R instances
(using foreach package for example). It uses `Redis` instance, a key-value store server <https://redis.io/>,  to follow progress of all the
tasks in a queue.
}
\author{
Clement Turbelin
}
