% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redisProgress.R
\docType{package}
\name{redisProgress}
\alias{redisProgress}
\alias{redisProgress-package}
\alias{redisProgress-package}
\title{Progress bar for distributed tasks with redis backend}
\description{
redisProgress Package
}
\author{
Clement Turbelin
}
\keyword{internal}
